genericapiserver: cut off pkg/serviceaccount dependency
